+++ src/Storage.Interface/Enums/ReferenceType.cs
@@ -1,8 +1,15 @@
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+using TextJson = System.Text.Json.Serialization;
+
 namespace Altinn.Platform.Storage.Interface.Enums;
 
 /// <summary>
 /// The type of the connected object
 /// </summary>
+[JsonConverter(typeof(StringEnumConverter))]
+[TextJson.JsonConverter(typeof(TextJson.JsonStringEnumConverter))]
 public enum ReferenceType
 {
     /// <summary>
