+++ src/Storage.Interface/Enums/RelationType.cs
@@ -1,8 +1,15 @@
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+using TextJson = System.Text.Json.Serialization;
+
 namespace Altinn.Platform.Storage.Interface.Enums;
 
 /// <summary>
 /// The type of relation to the connected object
 /// </summary>
+[JsonConverter(typeof(StringEnumConverter))]
+[TextJson.JsonConverter(typeof(TextJson.JsonStringEnumConverter))]
 public enum RelationType
 {
     /// <summary>
