+++ src/Storage.Interface/Models/SignDocument.cs
@@ -63,35 +63,5 @@ namespace Altinn.Platform.Storage.Interface.Models
             [JsonProperty(PropertyName = "signed")]
             public bool Signed { get; set; }
         }
-
-        /// <summary>
-        /// Information about the signee
-        /// </summary>
-        public class Signee
-        {
-            /// <summary>
-            /// The userId representing the signee.
-            /// </summary>
-            [JsonProperty(PropertyName = "userId")]
-            public string UserId { get; set; } = string.Empty;
-
-            /// <summary>
-            /// The partyId representing the signee.
-            /// </summary>
-            [JsonProperty(PropertyName = "partyId")]
-            public string PartyId { get; set; } = string.Empty;
-
-            /// <summary>
-            /// The personNumber representing the signee.
-            /// </summary>
-            [JsonProperty(PropertyName = "personNumber")]
-            public string? PersonNumber { get; set; }
-
-            /// <summary>
-            /// The organisationNumber representing the signee.
-            /// </summary>
-            [JsonProperty(PropertyName = "organisationNumber")]
-            public string? OrganisationNumber { get; set; }
-        }
     }
 }
